PR 81825: Merge users/jpricket/partial to master
These commits add a HelpPanel class to the Checkout and Import dialogs under the list of repos/projects for Team Services. The text asks the user if they "can't find what they are looking for". Clicking or hovering over this text reveals a Tooltip that has further instructions and some helpful links.
This workaround is to help the small number of users that run into this known limitation (the server not returning ALL accounts the user has access to) until changes on the server remove the problem.

Related work items: #518020