Merged PR 254713: Allow property change to show up as an edit for TFVC
Allow property change to show up as an edit for TFVC

GitHub issue opened that when you changed a file to be executable that the TFVC plugin couldn't handle it. looks like there was a status we weren't accounting for.  https://github.com/Microsoft/vso-intellij/issues/103

I tried to test other property changes but none of them returned a status change like was seen with changing the a file to be executable.