Set isModifiedDocumentTrackingRequired = false for the TFSChangeProvider (#269)
TFVC clients never pick up the unsaved changes from the editor anyway,
so this flag shouldn't be set to true in the first place. But it breaks
the new reactive client even more, because the client won't properly
pick up the changes after manual refresh (because otherwise the whole
scope will be considered as dirty and won't be revalidated during the
following save).